MiBombo_version1/requirements.txt
# ============================================
# MiBombo Station - Frozen Requirements
# Date: 2026-01-14
# ============================================
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coverage==7.13.1
cryptography==46.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
fonttools==4.61.1
h11==0.16.0
idna==3.11
influxdb-client==1.49.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mock==5.2.0
numpy==2.4.1
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
ply==3.11
psutil==7.2.1
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasyncore==1.0.5
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
pyotp==2.9.0
pyparsing==3.3.1
pysmi==1.6.2
pysmi-lextudio==1.4.3
pysnmp==4.4.12
pysnmpcrypto==0.0.4
pytest==9.0.2
pytest-cov==4.1.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.13.0
python-socketio==5.16.0
reactivex==4.1.0
reportlab==4.4.7
requests==2.32.5
scapy==2.7.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
